Refactored server api structure
