+++ internal/api/server/token_gen.go
@@ -1,0 +1,5 @@
+package server
+
+func GenerateToken(login, password string) string {
+	return login + password + "salt"
+}
